requirements.txt
python-socketio
websocket-client
requests
discord-webhook
configparser
matrix-nio
